requirements.txt
matplotlib==3.10.0
numpy==1.26.4
opencv-python==4.10.0
pandas==2.2.3
PyYAML==6.0.2
scikit-learn==1.6.1
seaborn==0.13.2
torch==2.7.1
torchvision==0.22.1
torchviz==0.0.3
plotly==6.0.0
